+++ src/main/java/com/example/mypetlife/controller/hospitalView/HospitalViewController.java
@@ -1,4 +1,4 @@
-package com.example.mypetlife.controller.view;
+package com.example.mypetlife.controller.hospitalView;
 
 import com.example.mypetlife.dto.hospital.DistrictResponseDto;
 import com.example.mypetlife.dto.hospital.gyeonggiDo.GyeonggiHospitalCountryResponseDto;
@@ -9,7 +9,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
